refactored data structures and types
