+++ limit_order_book/types/__init__.py
@@ -4,3 +4,8 @@ from types.order_id import OrderId
 from types.volume import Volume
 from types.order_side import OrderSide
 from types.ticker import Ticker
+
+from types.order import Order
+from types.order_without_order_id import OrderWithoutOrderId
+from types.trade import Trade
+from types.top_of_book import TopOfBook
