+++ tests/test_double_limit_order_book/test_double_limit_order_book_order_id_exists.py
@@ -10,8 +10,8 @@ from limit_order_book.types import IntPrice
 from limit_order_book.types import Volume
 from limit_order_book.types import OrderSide
 from limit_order_book.types import Ticker
-from limit_order_book.order import Order
-from limit_order_book.double_limit_order_book import DoubleLimitOrderBook
+from limit_order_book.types import Order
+from limit_order_book.data_structures.double_limit_order_book import DoubleLimitOrderBook
 
 
 def test_double_limit_order_book_order_id_exists():
