+++ tests/test_limit_order_book/test_limit_order_book_trade_update.py
@@ -4,11 +4,11 @@ from limit_order_book.types import IntPrice
 from limit_order_book.types import Volume
 from limit_order_book.types import OrderSide
 from limit_order_book.types import Ticker
-from limit_order_book.order import Order
-from limit_order_book.order_without_order_id import OrderWithoutOrderId
-from limit_order_book.trade import Trade
-from limit_order_book.top_of_book import TopOfBook
-from limit_order_book.limit_order_book_wrapper import LimitOrderBook
+from limit_order_book.types import Order
+from limit_order_book.types import OrderWithoutOrderId
+from limit_order_book.types import Trade
+from limit_order_book.types import TopOfBook
+from limit_order_book.limit_order_book import LimitOrderBook
 
 
 def test_limit_order_book_order_update():
