+++ tests/test_multi_ticker_limit_order_book/test_multi_ticker_limit_order_book_top_of_book.py
@@ -4,9 +4,9 @@ from limit_order_book.types import IntPrice
 from limit_order_book.types import Volume
 from limit_order_book.types import OrderSide
 from limit_order_book.types import Ticker
-from limit_order_book.order import Order
-from limit_order_book.top_of_book import TopOfBook
-from limit_order_book.multi_ticker_limit_order_book import MultiTickerLimitOrderBook
+from limit_order_book.types import Order
+from limit_order_book.types import TopOfBook
+from limit_order_book.data_structures.multi_ticker_limit_order_book import MultiTickerLimitOrderBook
 
 
 def test_multi_ticker_limit_order_book_top_of_book_buy_sequence():
