+++ tests/test_order_priority_queue/test_order_priority_queue.py
@@ -6,8 +6,8 @@ from limit_order_book.types import IntPrice
 from limit_order_book.types import Volume
 from limit_order_book.types import OrderSide
 from limit_order_book.types import Ticker
-from limit_order_book.order import Order
-from limit_order_book.order_priority_queue import OrderPriorityQueue
+from limit_order_book.types import Order
+from limit_order_book.data_structures.order_priority_queue import OrderPriorityQueue
 
 
 def test_order_priority_queue():
