+++ tests/test_single_side_limit_order_book/test_single_side_limit_order_book_trade.py
@@ -6,8 +6,8 @@ from limit_order_book.types import IntPrice
 from limit_order_book.types import Volume
 from limit_order_book.types import OrderSide
 from limit_order_book.types import Ticker
-from limit_order_book.order import Order
-from limit_order_book.single_side_limit_order_book import SingleSideLimitOrderBook
+from limit_order_book.types import Order
+from limit_order_book.data_structures.single_side_limit_order_book import SingleSideLimitOrderBook
 
 
 def test_single_side_limit_order_book_no_trade_no_previous_order():
