+++ tests/test_webserver/test_webserver_util.py
@@ -1,7 +1,7 @@
 
 from limit_order_book_webserver.convert_trades_to_fastapi_trades import convert_trades_to_fastapi_trades
 from limit_order_book_webserver.webserver import FastAPI_Trade
-from limit_order_book.trade import Trade
+from limit_order_book.types import Trade
 from limit_order_book.types import OrderId
 from limit_order_book.types import Ticker
 from limit_order_book.types import IntPrice
